refactor: move to api/v1 & dedicated file for server route
