+++ reverse_proxy/app/nginx/conf.d/locations/upload.inc
@@ -1,4 +1,4 @@
-location /api/upload { limit_req zone=flood burst=100 nodelay; proxy_pass
+location /api/v1/upload { limit_req zone=flood burst=100 nodelay; proxy_pass
 http://server:5000; proxy_redirect off; proxy_cache off; proxy_set_header Host
 $host; proxy_set_header X-Real-IP $remote_addr; proxy_set_header X-Forwarded-For
 $proxy_add_x_forwarded_for; proxy_set_header X-Forwarded-Host $server_name;
